Super fast algorithm
With the last algorithm, we were computing the possibilites many times per cell. Instead, we can directly propagate the constraint as soon as a new value is added to the soduko. This way we perform the minimal nuber of operation as we encounter new solved cells along the process of propagating constraints.

Deepcopy is also a lot cheaper because it is based only on arrays.

This results in roughly 7-8 times faster solving